HealthMetrix/requirements.txt
Flask
psutil
py-cpuinfo==9.0.0
GPUtil==1.4.0
pySMART==1.3.0
wmi
tqdm